# slices: make Clone preallocate and copy instead of solely invoking append