slices: make Clone preallocate and copy instead of solely invoking append

By preallocating a slice of equal size then invoking copy, this
change results in a speedup:

```
                  │   old.txt    │               new.txt                │
                  │    sec/op    │    sec/op     vs base                │
IntClone_10B-12     41.18n ±  1%   34.07n ±  1%  -17.26% (p=0.000 n=10)
IntClone_1Kb-12     719.5n ±  2%   718.9n ±  2%        ~ (p=0.912 n=10)
IntClone_10Kb-12    7.032µ ± 29%   6.946µ ± 14%        ~ (p=1.000 n=10)
IntClone_1Mb-12     239.6µ ±  1%   238.9µ ±  1%        ~ (p=0.353 n=10)
IntClone_10Mb-12    2.431m ±  1%   2.418m ±  1%        ~ (p=0.143 n=10)
ByteClone_10B-12    35.86n ±  0%   29.43n ±  0%  -17.92% (p=0.000 n=10)
ByteClone_10Kb-12   689.2n ±  6%   694.3n ±  3%        ~ (p=0.912 n=10)
geomean             3.138µ         2.964µ         -5.56%

                  │   old.txt    │                new.txt                │
                  │     B/op     │     B/op      vs base                 │
IntClone_10B-12       104.0 ± 0%     104.0 ± 0%       ~ (p=1.000 n=10) ¹
IntClone_1Kb-12     8.023Ki ± 0%   8.023Ki ± 0%       ~ (p=1.000 n=10) ¹
IntClone_10Kb-12    80.02Ki ± 0%   80.02Ki ± 0%       ~ (p=1.000 n=10) ¹
IntClone_1Mb-12     8.000Mi ± 0%   8.000Mi ± 0%       ~ (p=0.350 n=10)
IntClone_10Mb-12    80.00Mi ± 0%   80.00Mi ± 0%       ~ (p=0.173 n=10)
ByteClone_10B-12      40.00 ± 0%     40.00 ± 0%       ~ (p=1.000 n=10) ¹
ByteClone_10Kb-12   10.02Ki ± 0%   10.02Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean             28.97Ki        28.97Ki       -0.00%
¹ all samples are equal

                  │  old.txt   │               new.txt               │
                  │ allocs/op  │ allocs/op   vs base                 │
IntClone_10B-12     2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
IntClone_1Kb-12     2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
IntClone_10Kb-12    2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
IntClone_1Mb-12     2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
IntClone_10Mb-12    2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
ByteClone_10B-12    2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
ByteClone_10Kb-12   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean             2.000        2.000       +0.00%
¹ all samples are equal

```